refactoring to use DeviceMessageMetadata
