+++ osgp-adapter-domain-smartmetering/src/main/java/com/alliander/osgp/adapter/domain/smartmetering/application/services/NotificationService.java
@@ -13,9 +13,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.alliander.osgp.adapter.domain.smartmetering.application.mapping.NotificationMapper;
-import com.alliander.osgp.adapter.domain.smartmetering.infra.jms.core.DeviceMessageMetadata;
 import com.alliander.osgp.adapter.domain.smartmetering.infra.jms.ws.WebServiceResponseMessageSender;
 import com.alliander.osgp.dto.valueobjects.smartmetering.PushNotificationAlarm;
+import com.alliander.osgp.shared.helperobjects.DeviceMessageMetadata;
 import com.alliander.osgp.shared.infra.jms.ResponseMessage;
 import com.alliander.osgp.shared.infra.jms.ResponseMessageResultType;
 
@@ -48,6 +48,6 @@ public class NotificationService {
                 new ResponseMessage(deviceMessageMetadata.getCorrelationUid(), deviceMessageMetadata
                         .getOrganisationIdentification(), deviceMessageMetadata.getDeviceIdentification(),
                         ResponseMessageResultType.OK, null, pushNotificationAlarmDomain, deviceMessageMetadata
-                                .getMessagePriority()), deviceMessageMetadata.getMessageType());
+                        .getMessagePriority()), deviceMessageMetadata.getMessageType());
     }
 }
